+++ test/common/MemoryInterface.h
@@ -7,7 +7,7 @@ class MemoryInterface {
  public:
   // Write a value to memory at the given address.
   template <typename T>
-  void write_to_memory(const int baseAddress, const int index,
+  void write_to_memory(const long long baseAddress, const int index,
                        const float value, const int partition) {
     T typedValue = static_cast<T>(value);
     ac_int<T::width> bits = typedValue.bits_rep();
@@ -22,8 +22,9 @@ class MemoryInterface {
                           T::width / 8, bytes);
   }
 
-  virtual void write_bytes_to_memory(const int address, const int partition,
-                                     const int size, const char* bytes) = 0;
+  virtual void write_bytes_to_memory(const long long address,
+                                     const int partition, const int size,
+                                     const char* bytes) = 0;
 
   virtual ~MemoryInterface() = default;
 };
